// src/modules/fuel-station/dto/request/FuelType.enum.ts
export enum FuelTypeEnum {
  Biodiesel = "BIODIESEL",
  Bioetanol = "BIOETANOL",
  GasoleoA = "GASOLEO A",
  GasoleoB = "GASOLEO B",
  GasoleoPremium = "GASOLEO PREMIUM",
  Gasolina95E10 = "GASOLINA 95 E10",
  Gasolina95E5 = "GASOLINA 95 E5",
  Gasolina95E5Premium = "GASOLINA 95 E5 PREMIUM",
  Gasolina98E10 = "GASOLINA 98 E10",
  Gasolina98E5 = "GASOLINA 98 E5",
  Hidrogeno = "HIDROGENO",
}